requirements.txt
advocate==1.0.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiohttp-retry==2.8.3
aiosignal==1.3.1
APScheduler==3.10.4
asttokens==2.4.1
attrs==24.2.0
Authlib==1.3.2
babel==2.16.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
cachetools==5.5.0
calibreweb==0.6.23
certifi==2024.8.30
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.7
coverage==5.5
cryptography==43.0.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
docopt==0.6.2
EbookLib==0.18
executing==2.1.0
eyed3==0.9.7
fastjsonschema==2.20.0
filetype==1.2.0
Flask==3.0.3
flask-babel==4.0.0
Flask-Dance==7.1.0
Flask-HTTPAuth==4.8.0
Flask-Limiter==3.8.0
Flask-Principal==0.4.0
Flask-WTF==1.2.1
frozenlist==1.4.1
fsspec==2024.9.0
gcloud==0.18.3
gcsfs==2024.9.0.post1
google-api-core==2.20.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-cloud-bigquery==3.26.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.66.1
grpcio-status==1.66.1
httplib2==0.22.0
idna==3.10
ImageMagic==0.2.1
imgkit==1.2.3
importlib_resources==6.4.5
ipython==8.12.3
iso-639==0.4.5
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jwcrypto==1.5.6
libmagic==1.0
limits==3.13.0
loguru==0.7.2
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
multidict==6.1.0
mutagen==1.47.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
ndg-httpsclient==0.5.1
netifaces==0.11.0
netifaces-plus==0.12.2
numpy==2.1.1
oauth2client==4.1.3
oauthlib==3.2.2
ordered-set==4.1.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
paypalrestsdk==1.13.3
pdfminer==20191125
phonenumbers==8.13.48
pickleshare==0.7.5
pillow==10.4.0
pip-check-reqs==2.5.3
pipreqs==0.5.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==5.28.2
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.20.0
Pygments==2.18.0
PyJWT==2.9.0
PyMuPDF==1.24.11
PyMySQL==1.1.1
pyOpenSSL==24.2.1
pyotp==2.9.0
pyparsing==3.1.4
pypdf==4.2.0
PyPDF2==3.0.1
# pypiwin32==223
Pyrebase4==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jwt==4.1.0
python-magic==0.4.27
# python-magic-bin==0.4.14
pyttsx3==2.98
pytz==2024.2
# pywin32==308
pyzmq==26.2.0
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==0.10.1
rich==13.8.1
rpds-py==0.20.0
rsa==4.9
setuptools==75.1.0
six==1.16.0
soupsieve==2.6
# speechmatic @ file:///E:/work/calibre-web
SQLAlchemy==2.0.35
stack-data==0.6.3
stripe==10.12.0
tinycss2==1.3.0
toml==0.10.2
tornado==6.4.1
traitlets==5.14.3
twilio==9.3.2
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
Unidecode==1.3.8
urllib3==1.26.20
URLObject==2.4.3
Wand==0.6.13
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.4
# win32-setctime==1.1.0
wrapt==1.16.0
WTForms==3.1.2
yarg==0.1.9
yarl==1.13.1
